Create Wifi Speed Test
The intention with this project is to test and log the Wifi speed over a period of time to benchmark against what the ISP's advertised speeds are.  I wanted a log file that would be appended over time to display those results.  I also wanted to use Ookla speed test for the most accurate results.